Add files via upload
OnChange -The onChange event is often paired with the value attribute to create controlled components in React. This ensures that the component's state is always in sync with the UI

Event.Traget.Value- event.target.value is used to access the value of an input element (like a text box, dropdown, or checkbox) during an event, such as onChange, onInput, or onClick. It allows you to read and use the current value of the input field that triggered the event.